requirements.txt
agno==1.1.7
beautifulsoup4==4.13.3
langchain==0.3.19
langchain_community==0.3.18
langchain_core==0.3.40
langchain_elasticsearch==0.3.2
langchain_qdrant==0.2.0
numpy==2.2.3
python-dotenv==1.0.1
qdrant_client==1.13.2
sentence_transformers==3.4.1
streamlit==1.42.2
